Restructure of typography styles folder structure
