+++ src/styles/typography/body-1.tsx
@@ -1,5 +1,6 @@
-import { colors } from "../../colors";
-import { typography } from "../typography";
+import { colors } from "../colors";
+import { typography } from "./typography";
+
 
 export const Body1 = ({ children }: { children: React.ReactNode }) => {
     return (
