+++ src/styles/typography/dd.tsx
@@ -1,5 +1,5 @@
-import { colors } from "../../colors";
-import { typography } from "../typography";
+import { colors } from "../colors";
+import { typography } from "./typography";
 
 interface DTDDProps {
     children: React.ReactNode;
