+++ src/styles/typography/display.ts
@@ -1,5 +1,5 @@
-import { colors } from "../../colors";
-import { typography } from "../typography";
+import { colors } from "../colors";
+import { typography } from "./typography";
 
 export const DisplayStyle = {
   family: typography.font.family.primary,
