+++ src/styles/typography/h3.tsx
@@ -1,5 +1,5 @@
-import { colors } from "../../colors";
-import { typography } from "../typography";
+import { colors } from "../colors";
+import { typography } from "./typography";
 
 export const H3 = ({ children }: { children: React.ReactNode }) => {
     return (
