+++ src/styles/typography/h4.tsx
@@ -1,5 +1,5 @@
-import { colors } from "../../colors";
-import { typography } from "../typography";
+import { colors } from "../colors";
+import { typography } from "./typography";
 
 export const H4 = ({ children }: { children: React.ReactNode }) => {
     return (
