+++ src/styles/typography/style-customs.ts
@@ -1,5 +1,5 @@
-import { colors } from "../../colors";
-import { typography } from "../typography";
+import { colors } from "../colors";
+import { typography } from "./typography";
 
 export const customStyles = {
   caption: {
